Move back to opening output device when triggering
Previously we'd opened the output device at the start of the program.
This theoretically could reduce pops on playback. But unfortunately it
also means we may have ambient noise on the output device, and it will
use more power. Eventually PulseAudio can suspend the device, but we
don't want to rely on that.